repo-tools/github-version/requirements.txt
typer[all]
pydantic[dotenv]==1.*
PyGithub